Move app bar into widget
