+++ lib/views/feed.dart
@@ -2,15 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:snowpack/main.dart';
-import 'package:snowpack/models/aspect.dart';
-import 'package:snowpack/services/post_service.dart';
 import 'package:snowpack/views/post_list.dart';
-import 'package:snowpack/widgets/aspect_dropdown.dart';
-import 'package:snowpack/widgets/logo_button.dart';
-import 'package:snowpack/widgets/settings_button.dart';
-import 'package:snowpack/widgets/sort_button.dart';
-import 'package:snowpack/widgets/elevation_dropdown.dart';
-import 'package:snowpack/widgets/temperature_dropdown.dart';
+import 'package:snowpack/widgets/home_page_app_bar.dart';
 
 class Feed extends ConsumerStatefulWidget {
   const Feed({Key? key}) : super(key: key);
@@ -95,7 +88,10 @@ class _FeedState extends ConsumerState<Feed> {
           physics: const BouncingScrollPhysics(),
           controller: _scrollController,
           slivers: [
-            HomePageAppBar(context, colorScheme, postService),
+            HomePageAppBar(
+                context: context,
+                colorScheme: colorScheme,
+                postService: postService),
             CupertinoSliverRefreshControl(
               onRefresh: () => postService.refreshFeed(),
             ),
@@ -105,100 +101,4 @@ class _FeedState extends ConsumerState<Feed> {
       ),
     );
   }
-
-  SliverAppBar HomePageAppBar(
-      BuildContext context, ColorScheme colorScheme, PostService postService) {
-    void _showFilterMenu(BuildContext context, PostService postService) {
-      final initialElevationValue = postService.currentElevationFilter ?? 0;
-      final initialAspectValue = postService.currentAspectFilter;
-      final initialTemperatureValue = postService.currentTemperatureFilter ?? 0;
-
-      const aspects = Aspect.values;
-
-      showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            title: Text('Filter Posts'),
-            content: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                ElevationDropdown(
-                    postService: postService,
-                    initialElevationValue: initialElevationValue),
-                SizedBox(height: 20),
-                AspectDropdown(
-                    postService: postService,
-                    aspects: aspects,
-                    initialAspectValue: initialAspectValue),
-                SizedBox(height: 20),
-                TemperatureDropdown(
-                    postService: postService,
-                    initialTemperatureValue: initialTemperatureValue),
-              ],
-            ),
-            actions: <Widget>[
-              TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-                child: const Text('Close'),
-              ),
-              TextButton(
-                onPressed: () {
-                  postService.clearFilters();
-                  Navigator.of(context).pop();
-                },
-                child: const Text("Reset"),
-              ),
-            ],
-          );
-        },
-      );
-    }
-
-    void settingsPressed() {
-      print("Settings Pressed");
-      _showFilterMenu(context, postService);
-    }
-
-    return SliverAppBar(
-      floating: true,
-      flexibleSpace: Container(
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            colors: [colorScheme.primary, colorScheme.secondary],
-            begin: Alignment.topLeft,
-            end: Alignment.bottomRight,
-          ),
-        ),
-      ),
-      elevation: 4.0,
-      leading: LogoButton(context: context, colorScheme: colorScheme),
-      actions: [
-        SettingsButton(settingsPressed: settingsPressed),
-        SortButton(postService: postService),
-      ],
-    );
-  }
-
-  // DropdownButton<Aspect> AspectDropdown(PostService postService,
-  //     List<Aspect> aspects, Aspect? initialAspectValue) {
-  //   return DropdownButton<Aspect>(
-  //     hint: Text('Select Aspect'),
-  //     value: initialAspectValue,
-  //     onChanged: (value) {
-  //       if (value != null) {
-  //         postService.setAspectFilter(value);
-  //         print("Selected Aspect: ${value.toString().split('.').last}");
-  //       }
-  //     },
-  //     items: aspects.map((Aspect aspect) {
-  //       return DropdownMenuItem<Aspect>(
-  //         value: aspect,
-  //         child: Text(aspect.toString().split('.').last),
-  //       );
-  //     }).toList(),
-  //   );
-  // }
 }
